requirements.txt
Django>=5.0
gunicorn
psycopg2-binary
django-environ
django-redis
djangorestframework
django-apscheduler
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1